requirements.txt
# Core dependencies
streamlit>=1.22.0
crewai>=0.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.14.0

# Database
sqlite3-api>=0.1.0

# UI components
streamlit-extras>=0.2.0
streamlit-option-menu>=0.3.2

# AI and NLP
openai>=1.0.0
langchain>=0.0.267

# Utilities
pydantic>=2.0.0
python-slugify>=8.0.0